I.Helpr/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
mysql-connector-python==8.0.22
mysqlclient==2.0.1
protobuf==3.13.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
